requirements.txt
transformers>=4.43.0
datasets>=2.21.0
scikit-learn>=1.5.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
requests>=2.32.0
PyYAML>=6.0
bs4>=0.0.2              # meta‑package for BeautifulSoup4
captum>=0.7.0           # explainability (optional)
shap>=0.45.0            # explainability (optional)
protobuf>=4.25.3
tiktoken>=0.7.0
# requirements.txt
flask==2.3.2
torch>=1.10.0
transformers>=4.28.0
gunicorn==20.1.0

